backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psutil>=5.9.0
pywin32>=306
Pillow>=10.0.0
winsdk>=1.0.0b10
pyotp>=2.9.0
qrcode[pil]>=7.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
